fix levelUp:sp_Fw_F submit btn, backup db, admin del char with null/null-Transactions

+++ ppServer/backend/cards/signals.py
@@ -7,5 +7,7 @@ from .models import Card
 @receiver(pre_delete, sender=Card)
 def delete_transactions_when_deleting_card(sender, instance, **kwargs):
     ''' delete transactions that will be sender = receiver = None, therefore will get invalid '''
+    if instance.card_distributed_to_player:
+        raise ValueError(f"{instance._meta.object_name} kann nicht gelöscht werden. Sammle die physische Karte erst wieder vom Spieler ein.")
 
     instance.get_transactions().filter(Q(sender=None) | Q(receiver=None)).delete()
